+++ wow_dbc/CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * BREAKING: `Category` renamed to `ServerCategory`.
 * BREAKING: `Type` renamed to `OceanType`.
 * BREAKING: `ReputationRaceMask` renamed to `AllowedRace`.
+* BREAKING: `Flags` renamed to `PvpFlags`.
 
 ### Removed